fix: just one API call for ingredients data in App
