+++ src/pages/history.jsx
@@ -7,7 +7,6 @@ import Sidebar from '../components/sidebar/sidebar';
 import OrdersList from '../components/orders-list/orders-list';
 import Loader from '../components/loader/loader';
 // import slices and their functions
-import { getItems } from '../services/slices/items';
 import { getUser } from '../services/slices/user';
 import { getFeed } from '../services/slices/feed';
 
@@ -41,14 +40,6 @@ export const HistoryPage = () => {
   const userOrders = orders.filter((order) => (
     user.orders.includes(order.id)
   ));
-
-  // we need to have items from API in store to render ingredients icons in order card
-  useEffect(() => {
-    // won't call API if items are already in store
-    if (!itemsSuccess) {
-      dispatch(getItems());
-    }
-  }, [dispatch, itemsSuccess]);
 
   // we need to have user from API in store to select its orders
   useEffect(() => {
