+++ src/pages/ingredient.jsx
@@ -1,16 +1,11 @@
-import { useEffect } from 'react';
-import { useSelector, useDispatch } from 'react-redux';
+import { useSelector } from 'react-redux';
 import { useParams } from 'react-router-dom';
 // importing components from project
 import AppHeader from '../components/app-header/app-header';
 import IngredientDetails from '../components/ingredient-details/ingredient-details';
 import Loader from '../components/loader/loader';
-// import slices and their functions
-import { getItems } from '../services/slices/items';
 
 export const IngredientPage = () => {
-  const dispatch = useDispatch();
-  
   const {
     items,
     itemsRequest,
@@ -19,14 +14,6 @@ export const IngredientPage = () => {
   } = useSelector(
     state => state.items
   );
-
-  // we need to have items from API in store to render ingredient
-  useEffect(() => {
-    // won't call API if items are already in store
-    if (!itemsSuccess) {
-      dispatch(getItems());
-    }
-  }, [dispatch, itemsSuccess]);
 
   const currentIngredientId = useParams().id;
