+++ src/pages/order.jsx
@@ -6,7 +6,6 @@ import AppHeader from '../components/app-header/app-header';
 import Loader from '../components/loader/loader';
 import OrderDetailedView from '../components/order-detailed-view/order-detailed-view';
 // import slices and their functions
-import { getItems } from '../services/slices/items';
 import { getFeed } from '../services/slices/feed';
 
 export const OrderPage = () => {
@@ -27,14 +26,6 @@ export const OrderPage = () => {
   } = useSelector(
     state => state.feed
   );
-
-  // we need to have items from API in store to render ingredients icons in order card
-  useEffect(() => {
-    // won't call API if items are already in store
-    if (!itemsSuccess) {
-      dispatch(getItems());
-    }
-  }, [dispatch, itemsSuccess]);
 
   // we need to have feed from API in store to render order data
   useEffect(() => {
